PKG: move writeVersionFiles functionality to semanticVersion.py
This is because we can't correctly import semanticVersion from within writeVersionFiles when used in ascript (not package) due to Python's relative import issues
